+++ src/auth/mod.rs
@@ -2,3 +2,5 @@
 
 #[cfg(feature = "approle")]
 pub mod approle;
+#[cfg(feature = "token")]
+pub mod token;