+++ src/secrets/mod.rs
@@ -1,4 +1,6 @@
 //! Secret engine handles.
 
+#[cfg(feature = "kv1")]
+pub mod kv1;
 #[cfg(feature = "kv2")]
 pub mod kv2;